Fix autosplitter for Windows builds
Unfortunately, this means that this doesn't check if you're on the settings screen or in the game for the game state struct address.
This means that if you're still on the settings screen, it will try over and over to get the game state address even though it is not initialized yet.
I don't really know how to fix that for now so we'll just let it be